+++ packages/base/src/form-element.ts
@@ -16,25 +16,17 @@ limitations under the License.
 */
 
 import {addHasRemoveClass, BaseElement, CustomEventListener, EventType, SpecificEventListener} from './base-element.js';
+import {HTMLElementWithRipple, RippleInterface} from './utils.js';
 
 export {
   addHasRemoveClass,
   BaseElement,
   CustomEventListener,
   EventType,
-  SpecificEventListener
+  HTMLElementWithRipple,
+  RippleInterface,
+  SpecificEventListener,
 };
-
-export interface HTMLElementWithRipple extends HTMLElement {
-  ripple?: RippleInterface;
-}
-
-export interface RippleInterface {
-  activate: (e?: Event) => void;
-  deactivate: () => void;
-  handleFocus: () => void;
-  handleBlur: () => void;
-}
 
 export abstract class FormElement extends BaseElement {
   /**